chore: move react dependencies to preset and import them via bundler
The version numbers are copied from the target Liferay source tree (found
at the yarn.lock file).

The  preset would include all exported deps coming from Liferay, but I'm
only copying the ones needed (React), for demonstration purposes.

I add the deps to devDependencies so that they are available for building
and testing but are not included in the output JAR.

The imports config inside .npmbunderrc is copied from the global
npmscripts.config.js at Liferay source tree.

Since 7.4 GA2 we will be using "*" as imports semver expression which means
that any version coming from Liferay will do for the project. It's, thus,
the responsibility of the developer to make sure his code is valid for the
target platform.