chore: rely on bundler rules to copy static assets
Instead of using liferay-npm-build-support scripts (which, BTW, serve a
similar purpose to that of @liferay/npm-scripts) to copy static assets,
leverage bundler rules/loaders to do that task.

This is to show that -maybe- we can drive the whole build with the bundler
alone, not needing any external scripts or, at least, if we need them, we
can still reduce them to the minimum possible.