chore: rely on liferay-npm-bundler to invoke Babel
Again, this is a commit to leave the project's package.json as empty as
possible. Instead of invoking Babel from package.json we delegate that to
the bundler thus avoiding the need for any external script.

Note that we can apply this same strategy to other tasks like compiling
SOY, CSS, ... As I said, the benefits would be not needing any external
scripts at all.

Things worth to note:

1) The current Babel loader for npm-bundler only implements Babel 6, so
I've had to move to older devDependencies. This is, however, easy to fix,
as we only need to create a new loader for npm-bundler targetting Babel 7.

2) We have internalized .babelrc inside the preset. This is because the
current Babel loader for npm-bundler does not support .babelrc files, but
we can fix that too.