chore: final cleanup
This commit is not strictly necessary, but I wanted to see how empty the
package.json and .npmbundlerrc could be.

I'm removing the dependencies on liferay-npm-build-support so we cannot
translate or deploy the project any more if we apply this commit.

However, given that we will always be contained inside a blade project, I
would say we can/should delegate these tasks to the Java side, since they
are most suited to that side of the solution.

For testing this project in a real Liferay instance, just execute the
`build` script and manually copy the resulting JAR file (inside the
`dist` folder) to Liferay's deploy directory.